+++ src/SSMP2communication.cpp
@@ -531,7 +531,7 @@ void SSMP2communication::run()
 			op_str += "writeMulti_emulated_p";
 			break;
 		default:
-			op_str += "INVALID/UNKNOWN: " + QString::number(operation).toStdString();
+			op_str += "INVALID/UNKNOWN: " + QString::number( static_cast<int>(operation) ).toStdString();
 	}
 	std::cout << op_str << '\n';
 #endif
